+++ packages/Webkul/Webhook/src/DataGrids/LogsDataGrid.php
@@ -232,6 +232,19 @@ class LogsDataGrid extends DataGrid
      */
     public function prepareActions()
     {
+        if (bouncer()->hasPermission('configuration.webhook.logs.view')) {
+            $this->addAction([
+                'index'         => 'view',
+                'icon'          => 'icon-view',
+                'title'         => trans('webhook::app.configuration.webhook.logs.index.datagrid.view'),
+                'method'        => 'GET',
+                'frontend_view' => 'view-modal',
+                'url'           => function ($row) {
+                    return route('webhook.logs.show', $row->id);
+                },
+            ]);
+        }
+
         if (bouncer()->hasPermission('configuration.webhook.logs.delete')) {
             $this->addAction([
                 'index'  => 'delete',